MC-DATA : Created unit tests for VillagerType service and removed old test class
along the way, I made a few cleanups to toString methods, method and property types (Boolean object
to boolean primitive), and formatting.  Removed static main test methods

VillagerTypeService (just like the other service classes) has a method for find-or-create and a method
for create.  It seems to me now that this would allow the possibility of creating duplicates. I have
made some notes to reconsider the visibility of the create method.