MC-DATA : Created unit tests for Item service and removed old test class
ItemService had both a saveItem and an updateItem method.  These sound alike to me so I removed updateItem
Also added a logger to ItemDto to replace a sysout error message.
After commenting-out ServiceFactory several commits ago, I have now removed it

Made the usual cleanup changes:
* Boolean class to boolean primitive
* removed static main tests
* cleanup toString methods

While working on these tests, it occurred to me that getNextIdSeq is not an appropriate service
to be provided by the DAO classes.  I think no one outside the DAO would need to call this
service.  Only when saving a new object would the DAO need to derive a new ID sequence number
and there would be no need to expose the implementation of deriving this number.  I made comments
regarding this, but have not made code changes (yet).